Extract Creator-Kit file-browsing routes out of agent-channel.ts (Wave D batch 1)
agent-channel.ts 3311 -> 3110 lines. Moves KIT_FILES/KIT_SEARCH/
KIT_FILE/KIT_FILES_READ/KIT_FILE_FRAGMENT -- all thin wrappers over
kit-files.ts's pure functions -- into agent-channel-kit-files.ts as
registerAgentChannelKitFileRoutes(app, deps), following the same
register<Name>Routes(app, options) shape as creator-agent-key-routes.ts.

sendKitFilesError moves with the cluster (its only callers); kitFileStore
stays constructed in agent-channel.ts and is passed in as a dep since
the staging routes also need it. optionalFiniteQuery is duplicated
locally (cheap 3-line helper, still used by other routes in the parent
file).

Starts Wave D (agent-channel.ts decomposition, 3311 lines), the next
target after Wave B (submissions.ts) and Wave C (mcp-server.ts).
